reorganize files for demo
