+++ src/keri/base/__init__.py
@@ -4,12 +4,4 @@
 KERI
 keri.base package
 
-flo behaviors
-
 """
-import importlib
-
-_modules = ['resting', ]
-
-for m in _modules:
-    importlib.import_module(".{0}".format(m), package='keri.base')
